+++ webapp/src/main/webapp/iaf/gui/js/controllers.js
@@ -1727,7 +1727,12 @@ angular.module('iaf.beheerconsole')
 			});
 		}
 	};
-
+}])
+.controller('InlineStoreOverviewCtrl', ['$scope', 'Api', function($scope, Api) {
+	Api.Get("inlinestores/overview", function(data) {
+		$scope.result = data;
+	});
+	
 }])
 .controller('WebservicesCtrl', ['$scope', 'Api', 'Misc', function($scope, Api, Misc) {
 	$scope.rootURL = Misc.getServerPath();